[Essential]Refactoring tests
