+++ Essential_Feed/EssentialFeediOSTests/FeedUI/Helpers/UIButton+TestHelpers.swift
@@ -1,0 +1,7 @@
+import UIKit
+
+extension UIButton {
+    func simulateTap() {
+        simulate(event: .touchUpInside)
+    }
+}
