+++ Essential_Feed/EssentialFeediOSTests/FeedUI/Helpers/UIControl+TestHelpers.swift
@@ -1,0 +1,11 @@
+import UIKit
+
+extension UIControl {
+    func simulate(event: UIControl.Event) {
+        allTargets.forEach { target in
+            actions(forTarget: target, forControlEvent: event)?.forEach {
+                (target as NSObject).perform(Selector($0))
+            }
+        }
+    }
+}
