+++ Essential_Feed/EssentialFeediOSTests/FeedUI/Helpers/UIImage+TestHelpers.swift
@@ -1,0 +1,14 @@
+import UIKit
+
+extension UIImage {
+    static func make(with color: UIColor) -> UIImage {
+        let rect = CGRect(x: 0, y: 0, width: 1, height: 1)
+        UIGraphicsBeginImageContext(rect.size)
+        let context = UIGraphicsGetCurrentContext()!
+        context.setFillColor(color.cgColor)
+        context.fill(rect)
+        let img = UIGraphicsGetImageFromCurrentImageContext()
+        UIGraphicsEndImageContext()
+        return img!
+    }
+}
