+++ Essential_Feed/EssentialFeediOSTests/FeedUI/Helpers/UIRefreshControl+TestHelpers.swift
@@ -1,0 +1,7 @@
+import UIKit
+
+extension UIRefreshControl {
+    func simulatePullToRefresh() {
+        simulate(event: .valueChanged)
+    }
+}
